requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.44.1
crispy-bootstrap5==2025.6
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.7
django-cloudinary-storage==0.3.0
django-crispy-forms==2.4
django-environ==0.12.0
django-import-export==4.3.7
django-jazzmin==3.0.1
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
idna==3.11
lxml==6.0.2
openpyxl==3.1.5
psycopg==3.2.12
psycopg-binary==3.2.12
python-docx==1.1.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tablib==3.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
